pyproject.toml
[tool.poetry]
name = "git-grass-gardening"
version = "0.1.0"
description = ""
authors = ["hwang h <hehehwang@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
GitPython = "^3.1.24"
pyfiglet = "^0.8.post1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"
ipython = "^7.28.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
